requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-openai
langchain-pinecone
langchain-text-splitters
python-dotenv
tiktoken